requirements.txt
aiohttp>=3.8.0
asyncpg==0.29.0
flask-cors==6.0.2
flask==3.1.3
gunicorn==25.1.0
pgqueuer>=0.16.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
requests==2.32.5
psycopg2-binary
